🐛 option datatype
one issue, which wasn't very apparent, revolved around defining array of pins as u8 and not u32.

u8 makes sense, as the number of pins in any practical board is less than 255, yet, the rules here aren't made by me.

+++ src/config.rs
@@ -5,7 +5,7 @@ pub mod details {
 
     // LEDs
     pub const LED_COUNT: usize = 3;
-    pub const LED_PINS: [u8; LED_COUNT] = [
+    pub const LED_PINS: [u32; LED_COUNT] = [
         26, // 0  indicate power stability
         21, // 1  indicate internet connectivity
         23  // 2  indicate shift status
@@ -14,7 +14,7 @@ pub mod details {
 
     // Buttons
     pub const BUTTON_COUNT: usize = 5;
-    pub const BUTTON_PINS: [u8; BUTTON_COUNT] = [
+    pub const BUTTON_PINS: [u32; BUTTON_COUNT] = [
         25, // 0  terminate the program
         17, // 1  ports forwarded for virtual buttons
         27, // 2  ports forwarded for virtual buttons
@@ -25,7 +25,7 @@ pub mod details {
 
     // Buzzers
     pub const BUZZER_COUNT: usize = 1;
-    pub const BUZZER_PINS: [u8; BUZZER_COUNT] = [
+    pub const BUZZER_PINS: [u32; BUZZER_COUNT] = [
         13  // one pin
     ];
 }
